% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_reverb.R
\name{add_reverb}
\alias{add_reverb}
\title{Add a reverb effect to the audio signal, simulating a different recording environment}
\usage{
add_reverb(input_signal, type = "hall")
}
\arguments{
\item{input_signal}{numeric}

\item{type}{character, choice of reverb effect. Options are 'hall' and 'church'. Defaults to 'hall'}
}
\value{
numeric, vector representing the audio signal with the specified type of reverb applied.
}
\description{
Add a reverb effect to the audio signal, simulating a different recording environment
}
